Move to com.github namespace instead of de (#294)
