+++ plugin/src/test/fixtures/plugin_kotlin/src/main/resources/META-INF/gradle-plugins/com.github.triplet.play.test.properties
@@ -1,1 +1,1 @@
-implementation-class=de.triplet.gradle.play.test.AndroidDefaultConfigPlugin
+implementation-class=com.github.triplet.gradle.play.test.AndroidDefaultConfigPlugin
